# docs/api/openapi-v1.yaml
openapi: 3.1.0
info:
  title: Finanzas App API
  version: 1.0.0
  description: API v1 para app financiera Chile-first (CLP).
servers:
  - url: https://api.finanzasapp.cl/v1
security:
  - bearerAuth: []
...
paths:
...
  /transactions:
    get:
      tags: [Transactions]
      summary: Lista movimientos con filtros
      parameters:
        - in: query
          name: month
          schema:
            type: string
            pattern: '^\d{4}-\d{2}$'
        - in: query
          name: type
          schema:
            type: string
            enum: [income, expense]
        - in: query
          name: categoryId
          schema:
            type: string
            format: uuid
        - in: query
          name: source
          schema:
            type: string
            enum: [gmail, manual]
      responses:
        '200':
          description: Transacciones
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Transaction'
    post:
      tags: [Transactions]
      summary: Crear movimiento manual
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTransactionRequest'
      responses:
        '201':
          description: Creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transaction'
  /transactions/{transactionId}:
    patch:
      tags: [Transactions]
      summary: Editar categoria o glosa
      parameters:
        - in: path
          name: transactionId
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                categoryId:
                  type: string
                  format: uuid
                rawGlosa:
                  type: string
      responses:
        '200':
          description: Actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transaction'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Transaction:
      type: object
      properties:
        id: { type: string, format: uuid }
        type: { type: string, enum: [income, expense] }
        source: { type: string, enum: [gmail, manual] }
        amountClp: { type: integer }
        occurredAt: { type: string, format: date-time }
        rawGlosa: { type: string }
        merchant: { type: string }
        categoryId: { type: string, format: uuid }
    CreateTransactionRequest:
      type: object
      required: [type, amountClp, occurredAt, rawGlosa]
      properties:
        type: { type: string, enum: [income, expense] }
        amountClp: { type: integer, minimum: 1 }
        occurredAt: { type: string, format: date-time }
        rawGlosa: { type: string }
        categoryId: { type: string, format: uuid }
...